+++ CPanel/octreeFile.cpp
@@ -15,7 +15,7 @@ void octreeFile::writeFile(std::string filename,panelOctree* oct)
     std::vector<node<panel>*> nodes;
     nodes = oct->getNodes();
     
-    for (int i=0; i<nodes.size(); i++)
+    for (size_t i=0; i<nodes.size(); i++)
     {
         centers.push_back(nodes[i]->getOrigin());
         halfDs.push_back(nodes[i]->getHalfDimension());
@@ -26,11 +26,11 @@ void octreeFile::writeFile(std::string filename,panelOctree* oct)
     if (fid.is_open())
     {
         fid << (nodes.size()+1) << "\n";
-        for (int i=0; i<centers.size(); i++)
+        for (size_t i=0; i<centers.size(); i++)
         {
             fid << centers[i](0) << "\t" << centers[i](1) << "\t" << centers[i](2) << "\n";
         }
-        for (int i=0; i<halfDs.size(); i++)
+        for (size_t i=0; i<halfDs.size(); i++)
         {
             fid << halfDs[i](0) << "\t" << halfDs[i](1) << "\t" << halfDs[i](2) << "\n";
         }
@@ -49,7 +49,7 @@ void octreeFile::writeFile(std::string filename,particleOctree* oct)
     std::vector<node<particle>*> nodes;
     nodes = oct->getNodes();
     
-    for (int i=0; i<nodes.size(); i++)
+    for (size_t i=0; i<nodes.size(); i++)
     {
         centers.push_back(nodes[i]->getOrigin());
         halfDs.push_back(nodes[i]->getHalfDimension());
@@ -60,11 +60,11 @@ void octreeFile::writeFile(std::string filename,particleOctree* oct)
     if (fid.is_open())
     {
         fid << (nodes.size()+1) << "\n";
-        for (int i=0; i<centers.size(); i++)
+        for (size_t i=0; i<centers.size(); i++)
         {
             fid << centers[i](0) << "\t" << centers[i](1) << "\t" << centers[i](2) << "\n";
         }
-        for (int i=0; i<halfDs.size(); i++)
+        for (size_t i=0; i<halfDs.size(); i++)
         {
             fid << halfDs[i](0) << "\t" << halfDs[i](1) << "\t" << halfDs[i](2) << "\n";
         }
